requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/base.in requirements/dev.in
#
appdirs==1.4.4
    # via
    #   fs
    #   virtualenv
boto3==1.16.57
    # via fs-s3fs
botocore==1.19.57
    # via
    #   boto3
    #   s3transfer
click==7.1.2
    # via pip-tools
distlib==0.3.1
    # via virtualenv
django==2.2.17
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.in
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
fs-s3fs==1.1.1
    # via -r requirements/base.in
fs==2.4.12
    # via
    #   -r requirements/base.in
    #   fs-s3fs
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
packaging==20.8
    # via tox
pip-tools==5.5.0
    # via -r requirements/dev.in
pluggy==0.13.1
    # via tox
py==1.10.0
    # via tox
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.1
    # via botocore
pytz==2020.5
    # via
    #   django
    #   fs
s3transfer==0.3.4
    # via boto3
six==1.15.0
    # via
    #   fs
    #   fs-s3fs
    #   python-dateutil
    #   tox
    #   virtualenv
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via tox
tox-battery==0.6.1
    # via -r requirements/dev.in
tox==3.21.2
    # via
    #   -r requirements/dev.in
    #   tox-battery
urllib3==1.26.2
    # via botocore
virtualenv==20.4.0
    # via tox

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
